+++ CHANGES.rst
@@ -17,7 +17,7 @@ Changelog
 v4.4.1 (22 Dec 2019, @ankostis): bugfix debug print
 ===================================================
 + fix(net): had forgotten a debug-print on every operation call.
-+ doc(arch): explain :term:`parallel` & the need for :term:`marshalling <marshal>`
++ doc(arch): explain :term:`parallel` & the need for :term:`marshalling`
   with process pools.
 
 v4.4.0 (21 Dec 2019, @ankostis): RESCHEDULE for PARTIAL Outputs, on a per op basis